+++ crates/ln-dlc-node/src/tests/just_in_time_channel/create.rs
@@ -149,7 +149,7 @@ async fn fail_to_open_jit_channel_with_fee_rate_over_max() {
         )
         .unwrap();
 
-    payer.send_payment(&invoice).unwrap();
+    payer.pay_invoice(&invoice, None).unwrap();
 
     // Assert
 
@@ -203,7 +203,7 @@ async fn open_jit_channel_with_disconnected_payee() {
         )
         .unwrap();
 
-    payer.send_payment(&invoice).unwrap();
+    payer.pay_invoice(&invoice, None).unwrap();
 
     // We wait a little bit until reconnecting to simulate a pending JIT channel on the coordinator
     tokio::time::sleep(Duration::from_secs(5)).await;
@@ -289,7 +289,7 @@ pub(crate) async fn send_interceptable_payment(
         "Invoice amount larger than maximum inbound HTLC in payer-coordinator channel"
     );
 
-    payer.send_payment(&invoice)?;
+    payer.pay_invoice(&invoice, None)?;
 
     payee
         .wait_for_payment_claimed(invoice.payment_hash())
@@ -371,7 +371,7 @@ pub(crate) async fn send_payment(
         "Invoice amount larger than maximum inbound HTLC in payer-coordinator channel"
     );
 
-    payer.send_payment(&invoice)?;
+    payer.pay_invoice(&invoice, None)?;
 
     payee
         .wait_for_payment_claimed(invoice.payment_hash())