+++ crates/ln-dlc-node/src/tests/multi_hop_payment.rs
@@ -65,7 +65,7 @@ async fn multi_hop_payment() {
         invoice_amount_sat,
     );
 
-    payer.send_payment(&invoice).unwrap();
+    payer.pay_invoice(&invoice, None).unwrap();
 
     payee
         .wait_for_payment_claimed(invoice.payment_hash())