+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Contracts can only be entered in full amounts
 - New default amount is 10 contracts
 - Fix: Show pnl on success dialog when closing a position
+- Receive bitcoin in wallet
+- Send bitcoin in wallet to address or BIP21 URI
+- Redesign send and receive screens
 
 ## [1.4.1] - 2023-10-09